+++ src/main/java/woowacourse/auth/domain/user/privacy/Gender.java
@@ -1,4 +1,4 @@
-package woowacourse.auth.domain.user;
+package woowacourse.auth.domain.user.privacy;
 
 import java.util.Arrays;
 import java.util.NoSuchElementException;
