+++ src/main/java/woowacourse/auth/domain/user/privacy/Privacy.java
@@ -1,6 +1,7 @@
-package woowacourse.auth.domain.user;
+package woowacourse.auth.domain.user.privacy;
 
 import java.util.Objects;
+import woowacourse.auth.domain.user.address.FullAddress;
 
 public class Privacy {
     private final Name name;
