+++ src/test/java/woowacourse/auth/domain/user/AddressTest.java
@@ -7,6 +7,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import woowacourse.auth.domain.user.address.Address;
 import woowacourse.auth.exception.format.InvalidAddressFormatException;
 
 class AddressTest {
