+++ src/test/java/woowacourse/auth/domain/user/BirthDayTest.java
@@ -8,6 +8,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import woowacourse.auth.domain.user.privacy.BirthDay;
 
 class BirthDayTest {
