+++ src/test/java/woowacourse/auth/domain/user/DetailAddressTest.java
@@ -4,6 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import woowacourse.auth.domain.user.address.DetailAddress;
 
 class DetailAddressTest {
     @DisplayName("상세 주소 문자열을 전달받아 생성된다.")
