+++ src/test/java/woowacourse/auth/domain/user/GenderTest.java
@@ -10,6 +10,7 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.junit.jupiter.params.provider.ValueSource;
+import woowacourse.auth.domain.user.privacy.Gender;
 
 class GenderTest {
